+++ scripts/trim_galore.sh
@@ -25,6 +25,7 @@ if [[ $LG3_DEBUG ]]; then
   echo "- PWD=$PWD"
   echo "- USER=$USER"
   echo "- PBS_NUM_PPN=$PBS_NUM_PPN"
+  echo "- hostname=$(hostname)"
   echo "- TG=${TG:?}"
   echo "- CUTADAPT=${CUTADAPT:?}"
 fi